tasks/ml_dependency_injection/requirements.txt
numpy>=1.20.0
pandas>=1.3.0
pyyaml>=6.0 